-Refactored the oven classes to be cleaner and inherit from Vanillas BlockEntityOven.
-Moved handling of expanded food attributes (madeWith and expandedStats) to itemExpandedRawFood via the "OnBaked" implementation, removing a dependancy for ExpandedFood to be baked in our custom ovens.

-Modified our default config values and text to match vanilla values, as ours were by default slower than vanilla in all aspects.

These changes should allow other mods to bake ExpandedFoods natively, without custom compat behavior, and allow patching other modded ovens to work with ACA's config values, such as Decorative Ovens from Chisel Tools, so long as they dont require a custom BlockEntityClass and rely on Vanillas.
